// app/types/proparty.ts
export interface Iproparty {
    image1: string;
    image2: string;
    title: string;
    name: string;
    price: number;
    house1: string;
    house01: string;
    house2: string;
    house02: string;
    house3: string;
    house03: string;
    like: number;
    dislike: number;
    comment: string;
}